Use XR_FB_hand_tracking_aim to get trigger pinch status and factor
We are currently using our own algorithm to detect trigger button state
and pinching factor (a measure for distance between thumb and index
finger-tips) when in hand-tracking mode. We using this for all devices.

However, on Quest where XR_FB_hand_tracking_aim is supported, we can
leverage that extension instead. Note that Pico also supports the
extension, but unfortunately it doesn't give right values for pinching
status and strength.

This is specially important now for Quest, where a recent system upgrade
made the runtime hand-tracking more sensitive and our algorithm started
detecting repeated button trigger (pinches). Using the extension data
fixes these issues and is more future proof on these platforms.

As a bonus from this change, we move the pinch status and factor
calculation code to the GestureManager.