[Chromium] Add null checks for TextInput delegate
This is the exact same scenario as the one described in here
https://github.com/Igalia/wolvic/pull/951#issue-1881592422 but
for the TextInputImpl object. When the session is closed the
delegate is released, so any potential event coming from Chromium
will find that there is no delegate to notify.

Adding null checks to avoid doing any work in that case (and
prevent crashes).

Fixes #1001