Enable voice input from keyboard
We actually have -11 as the KEYCODE_VOICE_INPUT for inputting
using voice, but that keycode didn't get added to our custom
keyboard.

And since the keys are already some kind of full on the
keyboard, I instead choose to add a voice button to the left
of the keyboard.

I also clean up the original code a little bit so that we can
have the same experience as the voice search during voice input.
I also add new voice input prompt to differentiate that with
voice search.

Resolve #660

Signed-off-by: Songlin Jiang <sjiang@igalia.com>