[HVR] Do not request the WiFi SSID for mainland China packages
Requesting the SSID is considered a way to "collect user private information"
even if we only use it to show the WiFi SSID to the user in the tray (we
don't actually call home with those values nor anything).

We can prevent that call from happening for that specific context
(HVR builds for mainland China).