Make hint text visible in SingleEditSetting
Enable HintTextColor configuration in SingleEditSetting,
set the text color to grey so that it's visible to user.

e.g. It can be hard for user to learn that we will actually reset
the value in Homepage edit setting when we delete all the text
input there without seeing the hint text.

Signed-off-by: Songlin Jiang <sjiang@igalia.com>