Disable 1€ filtering for Magic Leap 2
We initially developed it for the Lenovo A3 which was not
providing stable and reliable joint data. As the results were
quite good we decided to use it for every single device.

However devices like the MagicLeap2 provide a quite
good and clean signal with almost zero noise, so let's better disable
it for those devices (their SDKs already do a similar filtering).

The GestureManager constructor has now a new parametter which is
a pointer to a struct with 3 floats. If that pointer is null then
we do no filtering. This is more complex than just using a boolean
but it'd allow us to provide per-device settings for the filtering.