Allow YouTube playing different projection types of VR videos continuously
Since YouTube plugin infers video projection types according to words
in video title and description, and if `mozVideoProjection` is already
there in the url, the plugin will respect the value and won't infer
any more.

This works perfectly fine for playing a single video, but if we try
to continue playing other projection types of video without leaving
the Youtube, the inferred `mozVideoProjection` will get preserved in the
url, which makes all the videos continue using the same projection type.

As a result, we should reinfer the page projection after navigating to a
new page that is different from initial page.

Also we should wait for yt-update-title event to happen so that we can
get the update title and description for projection inferring:
https://www.sitepoint.com/community/t/how-to-get-the-new-youtube-title-when-yt-navigate-finish-runs/377799/2

Remove the window load listener as this is not neccessary, youtube events
is enough.

Signed-off-by: Songlin Jiang <sjiang@igalia.com>