Fix voice search default language selection
We should not consider `default` as a language, as something
like `default_CN` will not be valid anyway, then it will
fallback to `en_US`. We should get the language from the default
local in this case.

We should also not directly use `Locale.getDefault()`, as e.g.
for Simplified Chinese, it will return `zh_CN_#Hans`, which will
fallback to `zh_HK`. We should reconstruct the defaultLocale and
extract only the language and country code so that it matches the
language list in mSupportedLocales

Signed-off-by: Songlin Jiang <sjiang@igalia.com>