Track hand-tracking joint radii in controllers
To support the WebXR Hand Input spec we need to keep track of the
radius of all joints in the controller object.
See <https://www.w3.org/TR/webxr-hand-input-1/#xrjointpose-interface>
Right now we only store the joint transforms.

Though we do store it in the OpenXR backend, the info is also needed
in common code to be able to move it to the web engine when system
state is pushed.

This change is part of an upcoming series that will add support for
WebXR hand-tracking in Wolvic/Chromium.