[OpenXR] Prefer BLEND_MODE_ALPHA over BLEND_MODE_ADDITIVE
Wolvic automatically selects the blend mode by inspecting the list
of available blend modes. In particular the order in which we select
the default blend mode is OPAQUE->ADDITIVE->ALPHA.

In the particular case of the MagicLeap2 that does not work well
because ADDITIVE was selected (as it does not support OPAQUE) and
that was causing issues with video playing. In particular audio
can be heard but no video was rendered. I guess the problem is
how video formats interact with an additive blend mode (in which
black pixels are considered transparent).

Everything works fine if we select ALPHA, so let's switch the
preferred order as it plays nicely with all of our supported
configurations.