[Chromium] Disable Vulkan for x64 builds
We've been seeing an important number of crashes on the Chromium
side with vulkan related backtraces when running it on a x64
device as the Magic Leap 2. Looks like Chromium automatically
enables Vulkan for x64. When using Chromium with ARM devices
the selected graphics context is OpenGLES instead (see
gpu/command_buffer/service/shared_context_state.cc in Chromium).

Furthermore it looks like it's interferring in some nasty ways,
for example videos do not render anything if vulkan is enabled.