Use move semantics to save hand transforms deep copies
When hand tracking is used we need to store the hand joint transforms
(matrices basically) in the controller containers. We do it for
every frame. Right now the set method receives a const reference
to a vector that is assigned to a vector member inside Controller.
This means that we are effectively doing deep copies of all those
vectors.

Instead we can just simply replace the const reference by a pass by
value attribute. Inside Controller we use move semantics to move
the function attribute to the class member. That way if the caller
passes a lvalue then the code does 1 copy, but if the caller passes
and rvalue (as this patch does using std::move) then no copy will
be made and vector internals are just being moved around.