Enable drag and move windows at X & Y direction
I also tried to enable the drag and move at Z direction but it
seems like we still have some issue with that, and it maybe okay
to not enable Z direction move as long as we have something like #734
landed.

Now we hardcode the max/min X/Y, later if we can change into spherical
move, we may then be able to remove that limit.

Partially resolve #927 and #99

Signed-off-by: Songlin Jiang <sjiang@igalia.com>