[HVR] Allow Wolvic to run in different phones with Huawei Glasses
Huawei Glasses can be attached to different Huawei phones. With
our new filtering based on the product model returned by Android,
we were limitting it to only run on the Mate 40 Pro (the one
we have been used for testing).

It's better not to be attached to any particular model in the case
of the HVR build to allow it to be run in multiple compatible
devices.