[Chromium] Fix emulated controller position
Right now Wolvic/Chromium shows a wrong position when controllers
are emulated during hand-tracking. This is due to the fact that blink
uses the grip space for the controller position, instead of the local
space.

For actual controllers, we already fixed this in 95c24cb48, and this is
the fix for emulated controllers. It sets the controller's
`transformMatrix` to the identity matrix, because this grip position's
origin coincides with the controller position. But then we need to
correct the start position and direction of the beam by setting the
controller's `beamTransformMatrix`; which is used together with the
`transformMatrix` to calculate origin and direction (see
Controller::StartPoint and Controller::Orientation).