Add support for Meta Quest 3 controllers (#1045)
Adding the required mappings and assets to display the Meta Quest 3
controllers in Wolvic. Not only that but we are now also able
to load the proper 3D models in WebXR sessions as well.

It requires a new patch that have just been added to the
wolvic-gecko repository adding support for this device
to Gecko.