Add check for Pico version >= 5.7.1
This is the version we know brought important changes (e.g, big
improvements to hand-tracking) so to play safe we are using this
version as the base one. But it is likely that the support for
XR_FB_passthrough was there in earlier versions.

We will update this version as we get more info from Pico.