Enable passthrough mode on Pico devices (#1066)
Now Pico runtime supports basic passthrough mode via the XR_FB_passthrough
extension <https://registry.khronos.org/OpenXR/specs/1.0/html/xrspec.html#XR_FB_passthrough>.

Since we already implement the extension, we just need to enable it for
Pico devices. The feature is enable for PicoOS >= 5.7.1

This is the version we know brought important changes (e.g, big
improvements to hand-tracking) so to play safe we are using this
version as the base one. But it is likely that the support for
XR_FB_passthrough was there in earlier versions.

We will update this version as we get more info from Pico.