Tooltip size independent of device density
To scale the size of tooltip widgets, we modify the "density" field
in the tooltip's placement.

Our implementation replaced the default density with a fixed value.
However, that default density is different between devices;
when we replace the device-dependent density value with a fixed number,
we introduce differences between devices.

In practice, this meant that tooltips could be too large
on some systems and too small on others.

This PR makes the tooltip's density dependent of the device's density,
so tooltips have the same relative size everywhere.