Scale down and change color of pointer target during select event
This patch extends the feature we currently have for pinch gesture,
where the pointer target is scaled as the pinch gesture closes, until
it is fully closed and the color is changed to blue.

With this, we provide a nice visual cue of the trigger event to the
user, and also make Wolvic interactions more consistent.