Use Khronos OpenXR headers in OCULUSVR builds
We have been traditionally using Oculus' OpenXR headers in the
OCULUSVR builds. However we don't use and we don't plan to use
any vendor specific API that is not part of the official spec.

So instead of using that header we just better stick to the
ones provided by the official repository.