[OpenXR] Add support for XR_FB_keyboard_tracking
Only Quest devices support this at the moment. Tested on Quest2.

The bulk of the implementation was added to OpenXRInput, bacause
OpenXRInputSource is meant for controllers and DeviceDelegateOpenXR
is already bloated. In the future, we might want to move this to
OpenXRTrackedKeyboard if code gets more complex.

Note that this patch doesn't deal with rendering of the keyboard, nor
the hands-on-keyboard passthrough functionality. It just track the
keyboard's id, availability and location on every frame. A subsequent
series will add the rest of the functionality to make this usable,
namely, XR_FB_render_model and XR_FB_passthrough_keyboard_hands
extensions support.