Remove UA MOBILE and use UA VR for mobile in settings
The current `User-Agent Mode` option in the Settings menu
is confusing.

It is a piece of code that we inherited from FxR and which
makes a very small change to the User-Agent when in Mobile
mode, but its practical application isn't clear, and it's
very unlikely that users will understand it.

Maybe a simpler and more user-friendly replacement could be
`Default User-Agent Mode`, with only "Desktop" and "Mobile"
as choices:

- "Desktop" -> `USER_AGENT_MODE_DESKTOP`
- "Mobile" -> `USER_AGENT_MODE_VR`

This change would also align with the UI feature that lets
the user toggle between desktop and mobile modes.

Signed-off-by: Songlin Jiang <sjiang@igalia.com>