Select 90Hz refresh rate for Quest3
We use the FB_display_refresh_rate extension to dynamically select
the device display refresh rate in those devices supporting the
extension. The extension just returns a list of available refresh
rates and we're the ones selecting one from the list of candidates.

To select one of them we return the first supported rate which is
equal or higher than a suggested refresh rate. That suggested
refresh rate is device specific. For the case of the Quest3 we
didn't have an specific value so the fallback one (60Hz) was
suggested. That's why on the Quest3 we were running at the
slowest supported refresh rate 72Hz.