[Chromium] Implement YouTube desktop mode (#1101)
Add a UA override for youtube.com so that chromium uses the
same UA as gecko. This ensures that the correct video modes are
available for 360 videos.

Add app=desktop to YouTube URLs. This change implements YoutubeUrlHelper 
class that automatically adds app=desktop query parameter to YouTube URLs. 
This functionality was implemented on the Chromium side before but didn't 
work correctly in some cases.