Increase logical size and resolution of Web pages
Adjust the default DPI that we pass to the Web engine so that the texture
for the default window will have twice the resolution.

Adjust the default density so that the default window has a logical size
of 1024x576 CSS pixels.

Set the size of the maximum window so that its associated texture fits
within 2560x2560, which is the maximum in some systems.

The window scaling will interpolate between the default and maximum values.

This commit needed to be reverted because it clashed with recent changes
related to motion event handling. See:
  https://github.com/Igalia/wolvic/pull/1105
  https://github.com/Igalia/wolvic/pull/1106
  https://github.com/Igalia/wolvic/pull/1107