[Chromium] Disable bitmap capture for tabs' view
Wolvic creates a bitmap for every tab with its contents to be displayed
in the tabs view. For that it uses the acquireDisplay() call. In the case
of Chromium that method incorrectly additionally performs some widget
manipulation (that should be done somewhere else). This makes the call
to captureBackgroundBipmap() dangerous because it will cause some
widgets to dissapear.

Let's disable capturing a bitmap until the above problem is not resolved
(it's actually hard). In any case it has never worked in Chromium
because it was not implemented yet.