Add a new callback parameter to WSession::onNewSession()
This new parameter will carry a callback that will be executed just
after creating the Session but before notifying the Session clients.
It will be eventually used by the Chromium port to perform some
extra initializations on the session. In particular this is needed
for the cases in which the session creation is triggered by the
web engine (like when navigating a target=_blank link).