[Chromium] Add support for target=_blank navigations
In these cases the renderer process creates new WebContents and
notifies the client (Wolvic) about that. However it is still
very difficult to accomodate that pattern with the current
Wolvic architecture which was mainly designed to fit well with
GeckoView API. In particular it'll be really hard to stack a
new Session and at the same time not creating a new TabImpl
with its own WebContents as it happens now. Needless to say,
that requires several new objects on the chromium side to
support that use case.

However we've found a different way to do that which does not
actually require changes in Chromium. It's based on a feature
of Chromium that allows embedders to create their own WebContents
even in those cases. If we declare that we will provide our own
WebContents (by returning false in shouldCreateWebContents()
called by WebContentsDelegateAndroid::IsWebContentsCreationOverridden)
and then later on we don't provide any (as the default implementation
of the delegate does) then Chromium considers that it
should not do any window creation. So Chromium does not
create any new window/contents but it has already notified
us about which was the target URL. Then we can proceed to
create a new session for that URL as usual in a new Tab.

Fixes #1003