Add head lock feature in hamburger menu
When head lock is enabled, the browser window follows you wherever you look.

It can be useful, according to user feedback, for wanting to watch tv shows while
stretching in a range of positions.

This can also help mitigate issue #927, where users can temporarily enable the mode
to rotate the window, so they can continue watching it easily when laying down, which
can be a good help with #1050

Signed-off-by: Songlin Jiang <sjiang@igalia.com>