Enhance detection of hand tracking availability
So far we've been using XrHandJointLocationsEXT::isActive to
determine whether or not hand tracking was active for a given
hand. However that is not a reliable way to detect that because
we've seen SDKs that return true for isActive when all the
hand joints poses are invalid.

That's why we added an extra computation to check that we have
at least one valid joint before considering that the hand
tracking is active for a given source.