Add a new VisionGlass build flavor and initial implementation
Vision Glass is a device from Huawei which consists on a see-passthrough
glasses that can be attached to a phone via an USB-C cable
https://consumer.huawei.com/cn/wearables/vision-glass/. They're
able to act as a "projector" of the phone contents. They're also
interesting for XR use cases because they have an IMU that is able
to report orientation data.

This PR implements the build support and a basic implementation.
So far we can display content on the glasses. There is no input
interaction yet (apart from touches on the phone screen that can be
used to generate input for Wolvic).