Support head lock
When head lock is enabled, the browser window remains in the center
of your vision.

This functionality can be enabled in the Display section in the Settings.

According to user feedback, this can be useful for watching video while
not in an upright position.

It will also be useful for devices with a very narrow field of view.

This can also help mitigate issue #927, where users can temporarily
enable the mode to rotate the window, so they can continue watching
it easily when laying down, which can be a good help with #1050

Based on work by Songlin Jiang <sjiang@igalia.com>

See also:
https://github.com/Igalia/wolvic/pull/1111

Signed-off-by: Felipe Erias <felipeerias@igalia.com>