Enable drag and move windows in X direction (#1050)
Windows can be moved just by dragging. This is only enabled
in the X direction and while in curved mode for the time being. We will
eventually extend it to the Y direction and the non-curved mode.

This conflicts with the head locking mode, that's why it isn't enabled
in that case.

Signed-off-by: Songlin Jiang <sjiang@igalia.com>
Co-authored-by: Felipe Erias <felipeerias@igalia.com>