MagicLeap2: Add 3D controller model
Adding the 3D model for the MagicLeap2 controller. Textures are
still not 100% correct but it's the best we could get from the
original blender file.

Also fixed the input mapping selection. We're providing the
generic KHR controller as input profile when emulating the
controller for hand input. Instead we should use the Magic
Leap 2 one.