[ML2] Disable Hardware Acceleration for rendering UI widgets
When hardware acceleration is used when rendering UI widgets in the
Magic Leap 2 devices several things go wrong:
* The app vanishes for 3-4 seconds waiting for a dequeueBuffer()
call to complete
* Several widgets are not updated although they are still working

That's why we're disabling hardware acceleration in the MagicLeap2
device until those issues are fixed in the platform. Apart from that
we hide the "enable hw acceleration" switch in developer options.