[VisionGlass] Render Wolvic in glasses and control UI on phone screen
This is a large change because it's difficult to split it in smaller
chunks as we're revamping the way Wolvic works in the VisionGlass.
In general this is not a good idea but provided that we're in early
stages of VisionGlass port this should be fine as there should not
be regressions as the previous funtionality was quite limited.

These are the main changes:
   * Wolvic is now rendered just in the external screen (the glasses)
     using Android's Presentation API.
   * The phone screen shows a different UI which consists on a quite
     minimalistic and simple UI that allows users to performs clicks
     on a touchpad and also a "back" button.

Apart from that there are some other changes:
   * Fix a crash with early setControllerOrientation call
   * Enable WebXR by properly setting EyeResolution
   * Enable clicks on WebXR experiences
   * Use the actual FOV of the device
   * Improved the "wait for USB permissions" code