Initialize the VR external context after initializing Java
The external context initialization should happen as soon as possible
because we need it to be available before setting the communication
channel between Gecko and Wolvic. That's why we used to do that
when creating BrowserWorld.

However that's too early because the external context initialization
require Java to be initialized. That's why we're moving the
initialization just after initializing java in the RenderContext.