Remove assert for unknown devices
An assert was added to the code that translates from Wolvic device
types to the web engine device types in order to quickly identify
those devices that were not properly setup in Wolvic.

However that assert is backfiring as it does not allow the
NoAPI flavour to launch WebXR experiences because the device
is very likely to be unknown (since it's meant to be run
in any android device like phones).

This PR first renames the preprocessor variable to make it
coherent with the rest of the flavours and then uses it to
prevent the NoAPI flavour from asserting.

We're also removing a redundant log message (the assert
does that job already).