Set the app_name value in build.gradle
Our AndroidManifest specifies that the app_name should be retrieved
from resources (by using android:label="@string/app_name"). That's
why we were setting the app_name in a non_l10n.xml resource file.
For debug builds we're using a different XML file that set the name
to "Wolvic (Dev)".

So far that approach has been working fine. However we'll have
devices that could have multiple versions of Wolvic installed.
For example a Huawei Mate 40 Pro phone could have the "traditional"
Wolvic for the Huawei Glasses and the "new" Wolvic for the
VisionGlass hardware. In those cases it'd be nice to have
different app names.

This means that we need a way to specify different app names
per flavour and on top of that, the possibility to append a
suffix ("Dev") for the debug builds.

We cannot do that with resource files because first we won't
have control on how the merges are done and also because we
could not correctly append debug suffixes unless we create a
resource file per each combination of build flavors.

The solution is to define the app_name as a resValue in the
build.gradle (this way we keep the resource available from
within the app) and then we can use applicationVariants.all
to append the debug suffix to debug builds. Additionally
we could eventually overwrite the app_name for specific
flavors (while keeping the option to add the debug suffix).