Suppress unchecked cast warning in CustomKeyboard
We do know that the rows field in Keyboard is an ArrayList<Row> so we can
safely remove the warning. Instead of suppressing it for the whole method
we can convert the if to a ternary operator and use the annotation for a
single line as it's done later in the same method.