Fix bug when moving large windows
When the user tries to move a large window (3x size), it can become
stuck at one extreme of the 3D environment.

The underlying cause seems to be a bug in how we compute the movement
of the pointer in extreme cases when the window widget is too large
and is positioned on a cylinder.

This bug is quite hard to pin down and only seems to happen when trying
to move a large window, so for the time being the fix is to restrict
the range of movement to the width of one default window.