[Pico] Rename PicoXR device type to Pico4x
The PicoXR name was selected by the first contributor to the Pico4 port
hoping to cover multiple devices with a single name. However the idea
behind DeviceType is precisely to add per-device specific code/settings.
For platform stuff we always have the different build flavours (in the
case of Pico we'll keep using PicoXR as it's now).

This will allow us to eventually support more OpenXR Pico devices,
like the Neo3.

Apart from the rename we're removing duplicated code like the input
profile for the Pico 4E which is exactly equal to the Pico 4. Actually
we don't have any Pico4E specific code (nor even the android model)
so it does make sense to merge both in the Pico4x device type.