[VisionGlass] Set the height of glasses and controller
The VisionGlass is a 3DoF device meaning that we don't get information
about the position of the device but just the orientation. This means
that in order to get a good default virtual position we need to force
the location of the device in the virtual world. For that we use
a fixed value which is the same that we have been using in the past
for other similar devices like the HVR.

Apart from that we're removing the code that we had for recentering
the view as it is not used and mainly inherited from the NoAPI flavour.