[VisionGlass] Use glasses' 3D rendering mode
VisionGlass API allows developers to request either a 2D operation
mode or a 3D one. In the case of 2D model the glass will project
the same image in both eyes. That's the mode that we've been
using as it was easier to setup.

In the case of the 3D operation mode, the left half of the viewport
will be projected only to the left eye, while the right half will
be used for the right eye. That allows us to produce slightly
different images for both eyes, and thus, generate the stereo
effect.

In order to make it work we had to basically implement the BindEye
DeviceDelegate method that in the case of the VisionGlass is used
to call glVieport to select the half of the texture were the draw
commands will operate. Note that as we have a single texture
we cannot call glClear in bindEye (as we do in the OpenXR backed
for example were we have an independent swapchain for each view/eye)
because that would clear the whole buffer. We have to do that
in startFrame().