Don't use the "desktop" viewport mode
In the "desktop" viewport mode, the Web engine renders pages inside
a special fixed-width viewport which is 980 CSS px wide.

Pages rendered on the "desktop" viewport mode keep their fixed width
when the containing window becomes larger, which is not very useful.

Furthermore, since we increased the screen density used by Wolvic,
this fixed width is actually smaller than our current default.

For these reasons, this PR removes usages of the "desktop" viewport
mode.

Note that "desktop mode" will still exist, but it will only affect
the User-Agent and not the viewport configuration.