Cleanup fullscreen code
The current code handling fullscreen is a mess and extremelly difficult
to follow. We've been adding patches over time to fix special situations
and that is making it even more unmaintanable.

The root problem is that there are two different fullscreen events
that are not properly handled. On the one hand we have the fullscreen
event from the content delegate. That could be triggered by any
HTML element that wants to go fullscreen (it does not have to be
media). And the other one is the fullscreen event triggered by media.
That is triggered after the previous one in the case of media
content.

Our code should perfectly handle the case of media fullscreen
while still allowing other HTML elements to go fullscreen as well.

This patch tries to address that by:
* splitting the code that performs the content fullscreen and the
media (video) fullscreen
* moving all the code related to media fullscreen to the handlers
of the media fullscreen event
* moving all the generic code that does window fullscreen to the
content fullscreen handlers. We know have enterFullscreenMode and
exitFullscreenMode for that.
* removing some handlers in the mAttachedWindow that are now
not needed.
* moving the creation of the projection menu to the media fullscreen
handlers as it is not required by content fullscreen