Fix controllers disappearing
This fixes a bug with video websites like https://www.ixigua.com/

The following sequence of steps:

- enter fullscreen mode
- leave fullscreen mode
- play the video

left the app in an unusable state because the controllers became hidden
and the user was not able to get them to appear again.

The cause was that when the page first enters fullscreen mode,
`NavigationBarWidget.onFullScreen` added a listener to the window
which would wait for the current video to become available
and would then update the UI for fullscreen mode.

That listener was not being removed when the page left fullscreen mode,
so when the user played the video later on, the listener would get
 activated and would try to transition the UI to fullscreen mode.
 This was leaving the app in an inconsistent state so clicks on the
 background were being ignored (`onWorldClick`).

 The fix is to ensure that the window listener is removed when the page
 leavs fullscreen mode.

 Just in case, `onWorldClick` will make the controllers visible by default.

 Fixes #1181