Handle URL bar editor commands in the UI thread
The OnEditorActionListener handler is not being executed in the UI thread.
This hasn't surprinsingly many issues so far but it's a real problem for
the Chromium backend because it will eventually invoke Tab's loadUrl().
The issue is that loadUrl() cannot be invoked from multiple threads
(it's normally invoked from the UI thread).