Support fullscreen vertical videos
Popular streaming services host more and more videos in vertical
format usually recorded with mobile phones. Wolvic was not properly
handling the case of watching them in fullscreen mode due to some
misshandling of some web events.

The confusion started because there are 2 different fullscreen
events coming from Gecko. On the one hand we have the content
fullscreen event that can be triggered by any HTML element that
goes fullscreen. And on the other hand there is the media fullscreen
event which is only triggered by media (video) elements.

When media goes fullscreen we typically get first the content
onfullscreen event and then the media onfullscreen event. But
the code that performs the window resize was only listening to
the content fullscreen event because that seemed enough. The
problem is that by the time we get the *content* fullscreen
event we don't usually have access to the media, and thus we
were only able to do a "normal" window resize to fullscreen
as we do with 2D movies for example. Then the *media* fullscreen
event was emitted with the media size information but that
was just ignored.

The solution is to indeed listen to the media fullscreen
event and perform the resize of the window there again so
that we could respect the video aspect ratio and also
the orientation.