[VisionGlass] Add a HOME button to the phone UI
Adding a home button that when pressed will behave exactly as the
home button in the navigation URL bar. In order to make the implementation
not very ugly we're adding a runPhoneUICallback() method that
performs the required checks to ensure that we access a valid
Windows object from the VRBrowserActivity subclass (yeah breaks
the usual OO model encapsulation).