Don't go back with buttons B and Y
We used to perform the go back action (or exit WebXR/immersive video)
with button B as it was used like that by some other apps/environments.
The truth is that nowadays almost nobody does that and what is worst,
it prevents WebXR apps from using the B button as it'll immediately
exit.

Fixes #1213