Update telemetry-related privacy options
We had a switch in the privacy options view to enable/disable sending
telemetry data to Mozilla. That was inherited from FxR and has not
been doing anything for a while. However it was still in the UI
which is confusing for users.

We're removing it except for the case of the HVR port as it has
an implementation using Huawei services. Users can still decide
whether to send that data or not.

Apart from that we're removing the crash reporting switch as we
are not sending any crash report anywhere. In the future it'd be
nice to collect that information but that requires some server
side infra that is not currently available/ready.