Synthesize FCP for cached pages
Our current code relies on the first-contentful-paint event to
among other things show the currently active window. That event
is neither emitted for pages loaded from back/forward cache nor
for pages completely cached by the HTTP cache.

This means that whenever we start Wolvic with a page that is
fully cached (like for example our start page) then Wolvic
will show a transparent window even if the page is properly
completelly loaded from disk.

To fix this we now listen to progress change events and trigger
the FCP if it was not triggered before.