Do not dump state of null sessions
Whenever a Session setups its listeners it calls dumpState to ensure
that no events are forgotten just because listeners might have been
setup too late.

As this happens very early it might be the case that there is no
session. If that's the case we should not call any event (actually
most of them are tagged with @NonNull which is not respected).