[VisionGlass] Use ElbowModel to compute the pointer transform
The position of the pointer did not properly match the position
of the beam although one followed the other. The problem is that
we were not properly computing the start and end positions of the
beam. In order to do that we can use the already existing
ElbowModel class which does compute it correctly.