Add a new handless ElbowModel
The ElbowModel in an utility class that allows us to compute a
beam based on a synthesized user's elbow position. This is
very useful when we have a 3DoF controller which tracks
orientation but not position.

Originally it was designed to generate a beam either for the left
or the right hand. However it's sometimes interesting to create
a neutral position in between both which would be valid for
both right-handed and left-handed people.