[VisionGlass] Use the largest window distance by default
Due to the constraints and FoV of the VisionGlass users cannot see
the whole browser window by default because it's too close. We're
modifying the settings so that the default window distance is the
larger available.

The whole window is not entirely visible though but it's the best
we can do ATM.