[VisionGlass] Remove call to SetBatteryLevel
That method is used to report controllers' battery levels to Java
so that their values could be shown in the tray bar. In the case
of the Vision Glass the controller is at the same time the VR
system (i.e. the phone) so it does not make sense to report any
value there.

The battery level of the phone is already retrieved from the
Java side.