Refresh battery levels less frequently
So far Wolvic has been refreshing the battery level of headset and
controllers with a 1s frequency. That's definitely way too much
because the change rate does not require that level of granularity.

What's more, the docs[1] mention "the impact of constantly monitoring
the battery level has a greater impact on the battery than your
app's normal behavior".

[1] https://developer.android.com/training/monitoring-device-state/battery-monitoring#MonitorLevel