[VisionGlass] Use 1€ filter to stabilize phone orientation
VisionGlass uses the phone as a controller. As phones have no grip
to grab them like XR controllers, the orientation data coming
from them tends to be "shaky" as it's very difficult to hold them
still.

That's why it's quite convenient to use the 1€ filter that we
use for hand tracking aim stabilization to "correct" the orientation
data coming from the phone accelerometer.