[VisionGlass] Do not rotate the Phone UI
Phone UI layout used to automatically switch to landscape mode when
the phone was rotated. That is an undesired capability for this
particular use case in which we want the UI to stay the same no
matter the rotation of the phone. That's because the phone is the
controller so we don't want the UI to suddenly change when the user
is pointing to some directions.