Copied links can be pasted on web forms
"Copy link" from the contextual menu adds the URL as data with
MIME type text/uri-list to the clipboard.

This data can be pasted on Android widgets like the URL bar,
but not on a form in a website.

The problem is that Gecko is only able to paste content of
type "text/plain". The solution is to store the link in both
formats (as URI and as text) in the clipboard.