Translated legal terms in Chinese
The legal terms for distribution in China are different from the
rest of the world.

The goal of this change is to decouple the Chinese release of Wolvic
from the translation of the app to the  Chinese language:

- packages for distribution in China will show the specific
legal terms for China regardless of language;
- in the rest of the world, the legal terms in the Chinese
language are a direct translation of our general terms.