Remove GoogleVR leftover
GoogleVR was supported by FxR at some point but then the support
was removed as it was not maintained. This PR removes some of
code that was added for GoogleVR but not removed.

As that code was deprecated we added a more modern version of it
as part of our effort to remove warnings in the code without
realizing that the code was added for an unsupported platform.
We're removing that code as well.