Introducing the PlatformActivityPlugin interface
The current architecture of the Activity class is suboptimal. The main
class, the VRBrowserActivity, turns out to be a subclass of a
PlatformActivity class which is platform specific (because some flavours
inherit from different types of Activity). The problem is that we cannot
add platform specifics to VRBrowserActivity, but at the same time it's the
class that has all the information, so it's complex to do anything
meaningful at the PlatformActivity level.

In order to do that in a clean way we're adding this new interface
which will be implemented and instantiated at the PlatformActivity
level. It'll contain the platform specific code, and as it will be
used by VRBrowserActivity, it would receive any required information
as parametters of the different interface methods.

This way we could encapsulate platform specifics at the PlatformActivity
level without losing access to the data/methods of the VRBrowserActivity.